documentation bug: NRPN flow doesn't match reality
reset was documented as cc#37,cc#36, which doesn't match ls_midi.ino:handleMidiInput().

The proper flow should be 99, 98, 6, 38, 99, 98 or (101,100 for issues with automation in some hosts)

I also added the decimal values, because after 30 years of this stuff, I still need to dig out a programmers calculator to convert back and forth from binary quickly.