Ugh.  pre-1.1.0-final merge.
Hopefully, I managed to bring my limiter along for the ride here without breaking anything. I moved the button to col 23 on global
because I can't tell just by skimming over this patch what the heck Geert is up to and he put in some stuff referencing (NUMCOLS-1)
so, we might be stepping on each other.  I moved it just in case.  That's if I managed to not screw this merge up. because git.
and endofunctors.

Merge branch 'master' of github.com:rogerlinndesign/linnstrument-firmware

Conflicts:
	linnstrument-firmware.ino
	ls_settings.ino