Fix for arpeggiator release bug.
When the arpeggiator is enabled and the active note is released, the arpeggiator pattern resets. This is described in detail in this thread:
https://www.kvraudio.com/forum/viewtopic.php?f=263&t=535623

Repro steps are:

1. Turn Arpeggiator on and set it to very slowly arpeggiate through a chord so it’s
easy to hear the problem. .
2. Hold a 4-note chord.
3. Immediately AFTER the second or third note in the arpeggio plays, release that
note.
4. Notice that a) the released note terminates immediately instead of sustaining
until the next note in the arpeggio, and b the arpeggio immediately resets to the
lowest note of the chord. It should continue to the next note in the chord. Notice
that this only occurs if you release your finger immediately AFTER its note plays
in the arpeggio and not if you release it immediately BEFORE it plays.

This PR resolves this issue by introducing the concept of a **stale note**. Before, if the active note in the arp sequence was released, the midi would be turned off and the arp sequence would be reset. Now, when the active note in the arp sequence is released, that note is marked as stale but the note remains in the cell mapping and the midi stays on.

Later on in `advanceArpeggiatorForSplit`, when the next note in the arp sequence is determined, we check to see if the previous note in the sequence is stale and remove it from the cell mapping at that time.

The variables I added to `NoteTouchMapping` increased its size by 2 bytes. Average free memory sits around 11,664 bytes, again about a 2 byte delta.